requirements.txt
streamlit
fastapi
uvicorn
pika
supabase
redis
python-dotenv
google-generativeai
aio_pika
streamlit-autorefresh
